NQ Vault update
Supports brute forcing of multiple galleries with different passcodes.
